workflows/docker-bake: support multi-platform images
Support building multi-platform images. For now, this means amd64 and
arm64, but the infrastructure allows for all kinds of images (yet GitHub
provides no suitable runners).

Note that this does *NOT* use qemu-emulation, as this takes a horribly
long time to execute for the type of full-system images we build.

Signed-off-by: David Rheinsberg <david@readahead.eu>